requirements.txt
torch==1.6.0
tqdm==4.48.2
click==7.1.2
pyyaml==5.1
